+++ src/parser/advancements/AdvancementHelper.js
@@ -2689,7 +2689,6 @@ Starting at 5th level, you can cast the ${lineageMatch.five} spell with this tra
 
     // You have resistance to all damage dealt by other creatures (their attacks, spells, and other effects).
     if (textDescription.includes("resistance to all damage dealt by other creatures")) {
-      grants.add("dr:all");
       Object.keys(CONFIG.DND5E.damageTypes).forEach((dr) => {
         grants.add(`dr:${dr}`);
       });